Relocate ApiVersions test helper
